requirements.txt
absl-py==1.0.0
appdirs==1.4.4
APScheduler==3.8.1
argon2-cffi==21.1.0
asgiref==3.3.1
asn1crypto==0.24.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
bleach==4.1.0
blinker==1.4
Brlapi==0.6.6
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
colorama==0.3.7
command-not-found==0.3
cryptography==2.1.4
cycler==0.11.0
dataclasses==0.8
decorator==5.1.0
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.3
Django==3.1.7
entrypoints==0.3
filelock==3.4.0
Flask==2.0.2
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.4
gast==0.5.3
google-api-core==2.2.2
google-api-python-client==2.29.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
grpcio==1.42.0
gunicorn==20.1.0
h5py==3.1.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.2
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.3
jsonschema==4.0.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
louis==3.5.0
Markdown==3.3.6
MarkupSafe==2.0.1
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.10.4
notebook==6.4.6
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
olefile==0.45.1
onboard==1.4.1
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.1.0
platformdirs==2.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.23
protobuf==3.19.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
Pygments==2.10.0
pygobject==3.26.1
pymongo==3.12.1
pyparsing==3.0.6
pyrsistent==0.18.0
python-apt===1.6.5-ubuntu0.6-elementary18-ubuntu5.1.7.1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.25
PyYAML==3.12
pyzmq==22.3.0
reportlab==3.4.0
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.8
SecretStorage==2.3.1
Send2Trash==1.8.0
six==1.16.0
SQLAlchemy==1.3.22
sqlparse==0.4.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==4.3.3
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.10.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.13.3
xkit==0.0.0
zipp==3.6.0
